+++ js/eme_admin_discounts.js
@@ -129,7 +129,7 @@ jQuery(document).ready(function ($) {
                 $('#DiscountsActionsButton').prop('disabled', true);
                 let ids = [];
                 selectedRows.each(function () {
-                    ids.push($(this).data('record')['id']);
+                    ids.push($(this).attr('data-record-key'));
                 });
 
                 let idsjoined = ids.join(); //will be such a string '2,5,7'
@@ -173,7 +173,7 @@ jQuery(document).ready(function ($) {
                 $('#DiscountHtoupsActionsButton').prop('disabled', true);
                 let ids = [];
                 selectedRows.each(function () {
-                    ids.push($(this).data('record')['id']);
+                    ids.push($(this).attr('data-record-key'));
                 });
 
                 let idsjoined = ids.join(); //will be such a string '2,5,7'