+++ js/eme_admin_holidays.js
@@ -46,7 +46,7 @@ jQuery(document).ready( function($) {
                     $('#HolidaysActionsButton').prop('disabled', true);
                     let ids = [];
                     selectedRows.each(function () {
-                        ids.push($(this).data('record')['id']);
+                        ids.push($(this).attr('data-record-key'));
                     });
 
                     let idsjoined = ids.join(); //will be such a string '2,5,7'