+++ js/eme_admin_members.js
@@ -297,7 +297,7 @@ jQuery(document).ready(function ($) {
                 $('#MembershipsActionsButton').prop('disabled', true);
                 let ids = [];
                 selectedRows.each(function () {
-                    ids.push($(this).data('record')['membership_id']);
+                    ids.push($(this).attr('data-record-key'));
                 });
 
                 let idsjoined = ids.join(); //will be such a string '2,5,7'
@@ -343,7 +343,7 @@ jQuery(document).ready(function ($) {
                 $('#MembersActionsButton').prop('disabled', true);
                 let ids = [];
                 selectedRows.each(function () {
-                    ids.push($(this).data('record')['members.member_id']);
+                    ids.push($(this).attr('data-record-key'));
                 });
 
                 let idsjoined = ids.join(); //will be such a string '2,5,7'
@@ -457,6 +457,8 @@ jQuery(document).ready(function ($) {
     // we add the on-click to the body and limit to the .eme_iban_button class, so that the iban-buttons that are only added via ajax are handled as well
     $('body').on('click', '.eme_iban_button', function(e) {
         e.preventDefault();
+        // clicking selects/deselects the row too, so invert it again
+        $('#MembersTableContainer').jtable('invertRowSelection',$(this).closest('tr'));
         let params = {
             'action': 'eme_get_payconiq_iban',
             'pg_pid': $(this).data('pg_pid'),