+++ js/eme_admin_rsvp.js
@@ -192,7 +192,7 @@ jQuery(document).ready(function ($) {
                             if (selectedRows.length > 0) {
                                 let ids = [];
                                 selectedRows.each(function () {
-                                    ids.push($(this).data('record')['booking_id']);
+                                    ids.push($(this).attr('data-record-key'));
                                 });
                                 let idsjoined = ids.join(); //will be such a string '2,5,7'
                                 $('.eme_jtable_button_for_pending_only .jtable-toolbar-item-text').text(emersvp.translate_pleasewait);
@@ -218,7 +218,7 @@ jQuery(document).ready(function ($) {
                             if (selectedRows.length > 0) {
                                 let ids = [];
                                 selectedRows.each(function () {
-                                    ids.push($(this).data('record')['booking_id']);
+                                    ids.push($(this).attr('data-record-key'));
                                 });
                                 let idsjoined = ids.join(); //will be such a string '2,5,7'
                                 $('.eme_jtable_button_for_approved_only .jtable-toolbar-item-text').text(emersvp.translate_pleasewait);
@@ -368,7 +368,7 @@ jQuery(document).ready(function ($) {
                 let ids = [];
                 let form;
                 selectedRows.each(function () {
-                    ids.push($(this).data('record')['booking_id']);
+                    ids.push($(this).attr('data-record-key'));
                 });
 
                 let idsjoined = ids.join(); //will be such a string '2,5,7'
@@ -440,6 +440,8 @@ jQuery(document).ready(function ($) {
     // we add the on-click to the body and limit to the .eme_iban_button class, so that the iban-buttons that are only added via ajax are handled as well
     $('body').on('click', '.eme_iban_button', function(e) {
         e.preventDefault();
+        // clicking selects/deselects the row too, so invert it again
+        $('#BookingsTableContainer').jtable('invertRowSelection',$(this).closest('tr'));
         let params = {
             'action': 'eme_get_payconiq_iban',
             'pg_pid': $(this).data('pg_pid'),