+++ js/eme_admin_templates.js
@@ -68,7 +68,7 @@ jQuery(document).ready( function($) {
                 $('#TemplatesActionsButton').text(emetemplates.translate_pleasewait);
                 let ids = [];
                 selectedRows.each(function () {
-                    ids.push($(this).data('record')['id']);
+                    ids.push($(this).attr('data-record-key'));
                 });
 
                 let idsjoined = ids.join(); //will be such a string '2,5,7'