+++ js/jtable/CHANGES.md
@@ -4,6 +4,7 @@ Short changes:
 * No need to hold CTRL-key for multisorting if it is active
 * Removed options: dialogShowEffect, dialogHideEffect
 * Removed deprecated private method `_addRowToTable` and rename `_addRow` to `_addRowToTable`
+* Added public methods deselectRow (Makes row/rows 'deselected') and invertRowSelection (Inverts selection state of a single row)
 * deleteAction, updateAction and selecting is now only possible if a key-field is defined (makes no sense otherwise)
 * Added option "roomForSortableIcon" (true/false) for sorting, so the sortable icon has room to appear next to the text
   True by default